Extract link resolution code from the documentation context (#306)
* Extract most link resolution code from DocumentationContext

* Stop waiting for request if external resolver client encounters fails

* Remove unnecessary calls to remove temp files in tests.

Also, fix a handful of typos in tests and comments.

* Add missing symbol relationships in test data

* Avoid changing kind of symbol in tests via known disambiguation

* Fix various correctness issues in test symbol graph files

* Fix inconsistent path component in test symbol graph files

* Minor grammar: 'indeterministic' -> 'non-deterministic'
