+++ Tests/SwiftDocCTests/Benchmark/TopicGraphHashTests.swift
@@ -1,7 +1,7 @@
 /*
  This source file is part of the Swift.org open source project
 
- Copyright (c) 2021 Apple Inc. and the Swift project authors
+ Copyright (c) 2021-2022 Apple Inc. and the Swift project authors
  Licensed under Apache License v2.0 with Runtime Library Exception
 
  See https://swift.org/LICENSE.txt for license information
@@ -82,7 +82,6 @@ class TopicGraphHashTests: XCTestCase {
         // Copy the test bundle and add external links to the MyKit Topics.
         let workspace = DocumentationWorkspace()
         let (tempURL, _, _) = try testBundleAndContext(copying: "TestBundle")
-        defer { try? FileManager.default.removeItem(at: tempURL) }
         
         try """
         # ``MyKit``
