+++ Tests/SwiftDocCUtilitiesTests/Test Bundles/TestBundle.docc/sidekit.symbols.json
@@ -77,7 +77,6 @@
         }
       ],
       "pathComponents" : [
-        "NonExistent",
         "UncuratedClass"
       ]
     },
